Adding debug logging for call request/response #344
The Call api endpoint isn't included in the Swagger
definition, so we hand build a http client. This doesn't include the
debug creation as that is performed by the Swagger Runtime code.

This change copies the logging approach from the Runtime code, adding
a context check for 'verbose' mode as well.